Forward agent containerd events to shim.
On a high-level, this commit:
1. Adds an EventExchange (existing containerd concept) to both the agent
   and shim. This is used by both for all event publishes
2. Creates a new concept, eventbridge.Getter, which is used to forward
   all events published on the agent's exchange to the shim's exchange.

The eventbridge.Getter is implemented as a TTRPC-based client/server
that provides just one API, GetEvent. When called from the client,
GetEvent will block until an event has been published on the server's
exchange, at which time that event will be returned. A separate
function, eventbridge.Attach, does that actual work of calling GetEvent
in a loop and forwarding the events to the local exchange.

This long-polling model is used because TTRPC does not support
streaming at this time.

Signed-off-by: Erik Sipsma <sipsma@amazon.com>