runtime: hard-code Firecracker API socket path
The Firecracker API socket is important for the runtime, as it uses the
socket to communicate with Firecracker.  Allowing the runtime to fully
control the API socket (including its path) removes a potential failure
condition from misconfiguration.

The API socket is hard-coded to exist within the current working
directory.  Part of the contract that containerd exposes for runtimes is
that they are started with the current working directory changed to be
that of the OCI bundle.  Using an API socket in the OCI bundle makes its
location well-known and predictable to the runtime.

Related: https://github.com/firecracker-microvm/firecracker-containerd/issues/59
Related: https://github.com/firecracker-microvm/firecracker-go-sdk/issues/88

Signed-off-by: Samuel Karp <skarp@amazon.com>