Add isolated e2e test for shim exit behavior.
This is the first e2e test that requires being run in isolation, as
being run on a host (or in a container) with multiple
firecracker-containerd instances would result in interference.

This thus adds a mechanism that allows buildkite to run each such
isolated test in its own docker container. The mechanism chosen here
is to name tests that require isolation with "_Isolated" at the end
and add a test util similar to the existing one for snapshotter root
tests that skips a test in an env var is not explicitly provided.

There is a script invoked as part of buildkite that finds all tests
with the "_Isolated" tag in their name and runs each one in its own
docker container.

This mechanism is minimally invasive to the actual test code itself
while not creating massive amounts of complication in our test infra
(such as other alternative methods like using build tags).

Signed-off-by: Erik Sipsma <sipsma@amazon.com>