Support specifying VMID when creating containers.
This is the first step for updating shims to support multiple containers
per-VM. It includes support for clients providing VMID and for the host
shim to parse that VMID, using it to decide whether a new shim needs to
spun up for the VM or if an existing one can be re-used.

Full support for multiple containers in a single VM requires more
changes in Agent; this change in particular just gives full support for
assigning single individual containers to different VMs.

Signed-off-by: Erik Sipsma <sipsma@amazon.com>