Generate both GRPC and TTRPC service definitions.
This will allow us to use the same proto definitions for both the external
facing API (which must be GRPC) and the internal facing API (which must be
ttrpc). For example, the FCControl plugin must accept GRPC requests from
clients, but now it can also make internal TTRPC requests to the shim using the
same API schema.

Signed-off-by: Erik Sipsma <sipsma@amazon.com>