Send shim logs through proxy.
Containerd's reference shim implementations just use a log FIFO provided by
and managed by containerd, present in the bundle dir for a task being managed
by the shim.

This doesn't work for us because our shim starts before bundle dirs exist and
can exist after tasks have exited.

The solution introduced here just has the shim's parent (the fccontrol plugin)
set the stdout/stderr of the shim process to point to containerd logs.

The downside of this solution is that if containerd restarts (restarting the
fccontrol plugin with it), then those stdout/stderr pipes will be closed and
the shim will need to restart to be able to log again.

If adding support for containerd restarts in the future, we will have options
to improve on this solution, such as creating+managing a FIFO for each shim from
the fccontrol plugin or updating containerd to do that for us (without being
tied to a bundle dir).

Signed-off-by: Erik Sipsma <sipsma@amazon.com>