Exit shim upon Task delete, not exit.
Issue #87 was previously addressed by having our shim process exit when all
tasks it was managing exited. However, it turns out that is not the expected
behavior for a containerd shim. The expected behavior is that a shim continue
running until all tasks have been *deleted* by the client (via the containerd
Delete API call).

When any task has been deleted (or create of a task fails) containerd will send
the shim a Shutdown API call, at which time the shim is expected to exit if it
*all* tasks it was managing have been deleted.

This commit fixes brings our shim in line with that expected behavior. It,
however, also retains the existing FCControl parameter "ExitAfterAllTasksGone"
(though now renamed to "ExitAfterAllTasksDeleted") to give users the ability
to specify whether the default containerd shim behavior should be ignored and
instead allow the Shim+VM to keep running even after all Tasks have been
deleted.

The majority of the changes here actually end up just being a refactor of the
TaskManager interface to safely handle Create/Delete task APIs alongside
checking whether the shim has any tasks left being managed (which is uses to
decide if it should shutdown). The refactorization also ensures that all IO
is flushed before a Delete call returns (which is a better solution to handling
I/O races than the temporary fix applied in 1e36219).

Signed-off-by: Erik Sipsma <sipsma@amazon.com>