Adding support for v0.17.0 of the SDK
This updates to use the newest version of the SDK which unconvered an
issue with one of the context that was passed during machine creation.
Originally it was using the request context which is a short lived
context and was causing the VM to be killed during execution. To fix
that the shim context is now used.

Signed-off-by: xibz <impactbchang@gmail.com>