Add performance test case comparing tc w/ bridge.
The intention of this test case is to add basic support for running performance
tests of various network configurations and to get a comparison of the CPU usage
differences between a TC-based VM network and a bridge-based VM network.

The test case added tries to make as fair a comparison as possible between TC
and a bridge. The TC approach uses the tc-redirect-tap CNI plugin, which results
in the following network layout for each VM:
[iperf server] <- vethA <- vethB <- [tc mirror] <- tap0 <- [iperf client]

In order to get a fair comparison, we want to run a bridge-based test with the
following layout:
[iperf server] <- vethA <- vethB <- [bridge] <- tap0 <- [iperf client]

Unfortunately, existing bridge based plugins usable in a general purpose test
like this cannot setup a bridge with a tap device like that, so this change
includes a barebones CNI plugin (under internal/cmd/test-bridged-tap/) that
sets up the network layout above w/ a bridge.

The test case then works by spinning up a configurable number of VMs, each of
which has a single container running an iperf client. The iperf client sends
data to a iperf server running on the host at a configurable bandwidth for a
configurable amount of time. The final output prints the average CPU cycles
consumed per second for the duration of the test case.

Signed-off-by: Erik Sipsma <sipsma@amazon.com>