Disable setting the shim as a subreaper and a reaper.
By default, containerd shim code results in our shim process being a subreaper
and running a reaper loop that waits on any children when SIGCHLD signals are
delivered. This does not work for us because we have code that uses "os/exec" to
spawn child processes (in the go sdk specifically). When "os/exec" waits on
child processes, it is in a race with the reaper loop, causing occasional
"wait: no child processes" errors when running commands.

Signed-off-by: Erik Sipsma <sipsma@amazon.com>