Chown newly created directories in jailer's ExposeDevice method.
Before this change, if exposing a device to a jail resulted in newly created
directories under the jail root, the directories were created with 0700 perms
but uid/gid owner set to root. This change results in the same 0700 perms but
chowns the new directories to the jail's uid/gid.

This is important as devices created by the devmapper snapshotter are multiple
levels underneath /dev, which means the jailer code now creates new directories
(as opposed to the naive snapshotter which didn't result in new directories
being created).

Signed-off-by: Erik Sipsma <sipsma@amazon.com>