Create a small Docker image for running integration tests
Building firecracker-containerd-integ-test takes time and
we cannot utilize build cache much since we copy the entire
source tree to the container image in the very beginning of
the build process.

This change introduces firecracker-containerd-integ-test-tiny
which doesn't build binaries such as firecracker-containerd
(they are built without Docker) but able to run "go test".

Signed-off-by: Kazuyoshi Kato <katokazu@amazon.com>