Build cni-bins in docker.
This is beneficial in that it removes another hard dependency on the host's
version of go. It also might prevent an issue I saw this morning in buildkite
where builds that previously passed started failing when compiling a third-party
dependency during the call to "cni-bins". The build was using the host's GOPATH
(not the docker volume gocache) and go version and I was only able to fix it by
manually logging into the host and clearing out the GOPATH directory.

Signed-off-by: Erik Sipsma <sipsma@amazon.com>