Re-establish a connection when vsock breaks
IOProxy struct connects the container's stdio inside a micro VM to
the FIFO files on its host. containerd clients consume the FIFO files
to access the container's stdio.

However, IOProxy was stopping itself when there was an error from
the streams, such as EPIPE. In that case, reading from the FIFO files
blocks indefinitely since there were no writers.

This change adds ioProxyHandler and uses the handler on State() to
make sure IOProxy is running. If not, ioProxyHandler#Attach() creates
a new IOProxy instance.

Fixes #482.

Signed-off-by: Kazuyoshi Kato <katokazu@amazon.com>