Fix kernel 5.10 config
Firecracker 1.0.0's build tool parses the kernel version
out of the specified config file to determine which kernel
sources to downlod.

The build tool requires the patch version to be specified:
https://github.com/firecracker-microvm/firecracker/blob/v1.0.0/tools/devtool#L2015

However, the linux kernel cdn drops the patch version if it is 0 (i.e.
sources for kernel 5.10.0 are hosted with a filename like
linux-5.10.tar.gz). This means that firecracker-1.0.0 cannot build
kernel 5.10.0.

Signed-off-by: Kern Walster <walster@amazon.com>