Replace address resolve with snapshot labels
Previously, the demux-snapshotter had to make an http request to the
http-address-resolver to find the vsock address and metrics
configuration for a remote snapshotter. With this change, the
http-address-resolver is removed and this information is sent to the
demux-snapshotter via snapshot labels directly from the client.

The labels are sent to the demux-snapshotter in the first pepare call
at which point the demux-snapshotter will establish connection with the
remote snapshotter. The connection is then cached via the namespace key
as it was in the previous implementation for subsequent calls.

Since each microVM houses its own remote snapshotter in a clean slate on
boot, any snapshot API call before a prepare will be operating on an empty
snapshotter and therefore will be a noop/not found. Therefore, if we
don't have a connection for an API, we can treat it as the noop/not found
case.

Signed-off-by: Kern Walster <walster@amazon.com>