Make agent static build the default for host build as well
When following the getting started guide, the commands are all executed on the
host and the agent built is dynamically linked. This will trigger an error when
running a container as the microVM image's glibc version will likely not match
the one from the host. Setting STATIC_BUILD to on for the host as is done for
the docker build ensure that the agent is suitable for use by the microVM.Ensure agent is build statically in all cases