Add a network namespace parameter to the CreateVM request
The current network namespace parameter nested in the JailerConfig
parameter requires specifying the entire jailer configuration, which is not
convenient and doesn't allow using the noop jailer with a network
namespace.

To overcome this limitation, add a new NetNS parameter to the CreateVM
request.

See also a similar change in firecracker-microvm/firecracker-go-sdk#155.

Signed-off-by: Georgiy Lebedev <lebedev.gk@phystech.edu>