Add self-contained kernel builder
Before the change, we relied on Firecracker's devtool to build a custom
kernel with our configs (which add FUSE support). The devtool's
build_kernel function is broken in Firecracker 1.4.1 and is completely
replaced by a non-containerized build script in main.

Rather than trying to get Firecracker's internal dev tooling to work for
our purposes, this change takes just the necessary scripts from
Firecracker to build a kernel in a container.

Signed-off-by: Kern Walster <walster@amazon.com>